Clean up header
