+++ src/components/Footer.js
@@ -8,7 +8,7 @@ const propTypes = {
 
 const defaultProps = {};
 
-class DefaultFooter extends Component {
+class Footer extends Component {
   render() {
 
     // eslint-disable-next-line
@@ -26,7 +26,7 @@ class DefaultFooter extends Component {
   }
 }
 
-DefaultFooter.propTypes = propTypes;
-DefaultFooter.defaultProps = defaultProps;
+Footer.propTypes = propTypes;
+Footer.defaultProps = defaultProps;
 
-export default DefaultFooter;
+export default Footer;
